requirements.txt
streamlit
pillow
easyocr
cryptography
base64
